Extracted subcomponent CustomTooltip
